User login, register & logout frontend and func
Added a new button for logging out.
Added a Register() function that calls specific API endpoint for registering users.
Added a Login() function that calls specific API endpoint for logging in.
Added a Logout() function that calls a specific API endpoint for logging out current logged in user.

Modified the variable(s) that is returned and binded with inputs. It is now an object.