Major overhaul on the search results
Made the results to be "cards" rather than just plain text
Added a add to playlist icon to "search-card"
Added further styling e.g. odd and even background colouring and hover colouring
Added a addToPlaylist() function that interacts with $store

Added a test button and function for checking the queue content. This shall be removed in production.